MotionDetector

Description

Module responsible for recognizing motion basing on accelerometer data

Documentation

data Motion

Represents user motion

Constructors

Undetected

Stopped

Running

Jumping

data MDState

State of detector

Constructors

MDState Motion MotionPosition Int Double Double MotionScaler

initMDState :: MDState

data Acceleration

Represents accelerometer data

Constructors

Acceleration Double Double Double Double Double Double

detectMotion :: MDState -> Acceleration -> (Motion, MDState)

Detects motion given detector state and acceleration Returns tuple of decected motion and new detector state